:fire: refactor: update packages, delete unused stuff
-Update json files and dependencies (react-router-dom)
-Move landingPage, pokemonCard, pokemonList and progress-bar into a
components folder
-Delete console.log and noscript tag from index.html & index.js
-Delete files from this directory and move to components

PKM-2
