+++ public/index.html
@@ -34,7 +34,7 @@
     <link rel="stylesheet" href="../src/css/styles.css" type="text/css">
   </head>
   <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <!-- <noscript>You need to enable JavaScript to run this app.</noscript> -->
     <div id="root"></div>
     <!--
       This HTML file is a template.
