+++ src/components/pokemonList.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const PokemonList = () => {
+  return (
+    <>
+      <h1>Pokemon List will go here</h1>
+    </>
+  )
+}
+
+export default PokemonList;
