+++ src/components/progress-bar.js
